+++ Runtime/Attributes/LogCalledAttribute.cs
@@ -2,7 +2,7 @@
 
 namespace Hertzole.CecilAttributes
 {
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, AllowMultiple = false)]
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, AllowMultiple = false, Inherited = false)]
     public sealed class LogCalledAttribute : Attribute
     {
         private bool logPropertyGet;