+++ Runtime/Attributes/MarkInProfilerAttribute.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace Hertzole.CecilAttributes
+{
+    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false, Inherited = false)]
+    public sealed class MarkInProfilerAttribute : Attribute { }
+}