+++ Runtime/Attributes/TimedAttribute.cs
@@ -2,6 +2,6 @@
 
 namespace Hertzole.CecilAttributes
 {
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, AllowMultiple = false)]
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, AllowMultiple = false, Inherited = false)]
     public sealed class TimedAttribute : Attribute { }
 }